Tests/LibWeb/Text/expected/content-blocker-shadow-pseudo-class-invalidation.txt

hover before: block
hover after enter: none
hover after leave: block
